イメージパスが存在しないバイクの場合を想定して、判定を追加。

+++ app/Http/Controllers/MotorcycleController.php
@@ -95,7 +95,9 @@ class MotorcycleController extends Controller
     {
         $motorcycle = Motorcycle::findOrFail($id);
 
-        Storage::disk('public')->delete($motorcycle->image_path);
+        if($motorcycle->image_path) {
+            Storage::disk('public')->delete($motorcycle->image_path);
+        }
 
         $motorcycle->delete();
